Move KFP SageMaker test  (#618)
**Which issue is resolved by this Pull Request:**
Resolves #

**Description of your changes:**
KFP SageMaker test is unnecessary in port_forward_sanity and causing
flakey issues, while we diagonse those will move it to test_sanity.py

**Testing:**
- [ ] Unit tests pass
- [ ] e2e tests pass
- Details about new tests (If this PR adds a new feature)
- Details about any manual tests performed

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.
